Add python-M2Crypto dependency.
This is required for certain usages of upstream salt modules (x509
is a good example). We are trying to keep this image as generic as
possible, but it's good to include dependencies that upstream
modules might need so it's straightforward to use this image as a
base for your own deployments.