portus: provide a flag for the background process
The background process has been introduced into portus, and it is way
easier to simply have an environment variable that toggles this
functionality.

The `PORTUS_INIT_COMMAND` environment variable is left untouched for
debugging purposes.

Moreover, this commit also updates the GEM_PATH environment so it uses
ruby 2.4 and not 2.1.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>