portus: added the devel:languages:ruby repo
In order to keep Virtualization:containers:Portus cleaner, we have removed some
packages from there and we are fetching them now from devel:languages:ruby. I've
changed the code so the GPG key for this repo is handled as well.

Moreover, this commit also contains some needed changes on the init file as for
the migration to ruby 2.6.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>